Revert "10 feature oauth client"
