--- credentials/credential.go
@@ -1,4 +1,0 @@
-package credentials
-
-type Credential struct {
-}
